requirements.txt
Flask==2.3.2
gunicorn==20.1.0
PyJWT==2.8.0
